requirements.txt
beautifulsoup4==4.12.3
camelot-py==0.11.0
ghostscript==0.7
lxml==5.3.0
opencv-python==4.10.0.84
pandas==2.2.2
polars==1.6.0
requests==2.32.3
rich==13.8.1
tk==0.1.0